Alert moved to common and fixed a small bug
