+++ common-ui/interfaces/alert.ts
@@ -1,0 +1,6 @@
+export type Alert = {
+  code: number | string;
+  message: string;
+  visible?: boolean;
+  displayText: string;
+};
